flink/jobs/requirements.txt
apache-flink==2.0.0
pillow
numpy
scikit-learn
scipy

